refactor: 툴팁 마크업 위치 이동 툴팁이 뜨는 기준인 roomTitle이 SettingRoom 컴포넌트에서 쓰이지 않아서 툴팁과 roomTitle 모두 SettingBar로 이동
