+++ client/src/components/study/SettingRoom/SettingRoom.module.css
@@ -6,39 +6,19 @@
   background-color: var(--gray950);
 }
 
-.video_component {
+.video_container {
   flex-grow: 1;
   display: flex;
 }
 
-.setting_tooltip,
-.plan_tooltip {
-  position: absolute;
-  top: -70px;
-  left: 16px;
-}
-
-.plan_tooltip {
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.bar {
-  position: relative;
-}
-
-.video_container {
+.video_wrapper {
   width: 100%;
   display: flex;
   justify-content: center;
-}
-
-.video {
-  height: 100%;
-  object-fit: cover;
-  transform: rotateY(180deg);
-}
-
-.video_wrapper {
   position: relative;
+}
+
+.video_wrapper video {
+  height: 100%;
+  transform: rotateY(180deg);
 }
