+++ client/src/components/study/StudyBar/StudyBar.module.css
@@ -1,6 +1,7 @@
 .bar {
   display: flex;
   align-items: center;
+  width: 100%;
   height: 10vh;
   background-color: var(--gray950);
   color: var(--gray300);
